Required Dataset
This is the dataset that is used for working with this project for analysing user data and video recommendations in streamlit app. 

NOTE: Only this dataset is applicable for implementing this project as the project code contains only this dataset columns value.

For further generalization, you can download this dataset and modify the values present in the dataset and implement it with the project.