Working model code with output
This is the file that is implemented and executed in google colab notebook. This file consists of all key aspects involved with the project like: Data preprocessing, K-Means clustering segmentation, LSTM model  building and training, user behavior analysis and future predictions based on video and user ids. 

You can use the working logic of the project from this file and implement it in google colab or jupyter notebook for predictions.